Switch to dynamic array for transitions, inline graph visits.
Since `Nfa` states are added sequentially and never removed, the number
of elements in the `transitions` field will always be equal to the
id of the largest element: this means that the map is dense, and a
simple dynamic array will be faster and consume less memory.

The functions that made use of a `visit` recursive subroutine
have been switched to an iterative version, which should not run
the risk of busting the stack.