Fix a bug that caused incorrect matches.
the regex `"a?"` used to match the string `"b"` due to the `match` function
incorrectly consuming the `'b'` rune in the `Epsilon` transition.

For the future, I would propose storing the `Epsilon` transitions from
the rest, since they are conceptually different and require quite a bit
of special logic.

this patch also contains a compilation pass that keeps all epsilon
transitions at the beginning of the array, as well as modified logic
in `match` to exploit this invariant and skip unnecessary work.